+++ packages/onchainkit/src/OnchainKitProvider.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { setOnchainKitConfig } from '@/core/OnchainKitConfig';
-import { useLayoutEffect, useMemo } from 'react';
+import { useContext, useEffect, useLayoutEffect, useMemo } from 'react';
 import { DefaultOnchainKitProviders } from './DefaultOnchainKitProviders';
 import OnchainKitProviderBoundary from './OnchainKitProviderBoundary';
 import { DEFAULT_PRIVACY_URL, DEFAULT_TERMS_URL } from './core/constants';
@@ -10,6 +10,8 @@ import type { OnchainKitProviderReact } from './types';
 import { generateUUIDWithInsecureFallback } from './utils/crypto';
 import { OnchainKitContext } from './useOnchainKit';
 import { useTheme } from './internal/hooks/useTheme';
+import { clientMetaManager } from './core/clientMeta/clientMetaManager';
+import { MiniKitContext } from './minikit/MiniKitProvider';
 
 /**
  * Provides the OnchainKit React Context to the app.
@@ -35,6 +37,12 @@ export function OnchainKitProvider({
   useLayoutEffect(() => {
     document.documentElement.setAttribute('data-ock-theme', theme);
   }, [theme]);
+  const isMiniKit = !!useContext(MiniKitContext)?.__isMiniKit;
+
+  useEffect(() => {
+    if (clientMetaManager.isInitialized()) return;
+    clientMetaManager.init({ isMiniKit });
+  }, [isMiniKit]);
 
   // eslint-disable-next-line complexity
   const value = useMemo(() => {