+++ packages/onchainkit/src/index.ts
@@ -16,5 +16,6 @@ export type {
   OnchainKitConfig,
   OnchainKitContextType,
 } from './core/types';
+import './styles/index.css';
 
 export { Connected } from './connected';